pyproject.toml
[project]
name = "arm-03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch>=9.2.0",
    "fastapi>=0.123.4",
    "pydantic-settings>=2.12.0",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
]
